move relevant useAuthStatus functions to identity module
